Add files via upload
Multiple Linear Regression (MLR) is an extension of simple linear regression where one dependent variable is predicted using two or more independent variables. Here, a synthetic regression dataset is generated for 2 features and 1 target(dependent) variable for explaining how data is fitted and model is trained using the in-built sklearn package and associated libraries .Later, the model is tested and is evaluated using the performance matrices -MAE,MSE and R2 Score.